# public/doc/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Disputes API",
        "version": "0.1"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "login",
                "operationId": "login",
                "requestBody": {
                    "request": "User",
                    "description": "User object that needs to be login or if not exist create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC8wLjAuMC4wOjgwODhcL2F1dGhcL2xvZ2luIiwiaWF0IjoxNjIwMjkzODM5LCJleHAiOjE2MjAyOTc0MzksIm5iZiI6MTYyMDI5MzgzOSwianRpIjoiUG4yOHhPMHo4VHZLRjBzVyIsInN1YiI6MTIzLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.fsfFD7PQ3MePJz7TIfW6pXHec71ZjgE2laoyzrWJgv0"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "me",
                "description": "Return user data",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "Authorization": [
                            "Authorization:Bearer token"
                        ]
                    }
                ]
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "logout",
                "description": "logout user, delete token",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "Authorization": [
                            "Authorization:Bearer token"
                        ]
                    }
                ]
            }
        },
        "/auth/refresh": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "refresh",
                "description": "refresh token",
                "operationId": "refreshToken",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC8wLjAuMC4wOjgwODhcL2F1dGhcL2xvZ2luIiwiaWF0IjoxNjIwMjkzODM5LCJleHAiOjE2MjAyOTc0MzksIm5iZiI6MTYyMDI5MzgzOSwianRpIjoiUG4yOHhPMHo4VHZLRjBzVyIsInN1YiI6MTIzLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.fsfFD7PQ3MePJz7TIfW6pXHec71ZjgE2laoyzrWJgv0"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "Authorization": [
                            "Authorization:Bearer token"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "uuid": {
                        "title": "uuid",
                        "type": "integer",
                        "example": 123456
                    },
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "datetime",
                        "example": "2021-05-03T16:29:10.000000Z"
                    },
                    "updated_at": {
                        "type": "datetime",
                        "example": "2021-05-03T16:29:10.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "tags": [
        {
            "name": "auth",
            "description": "authentication in app"
        }
    ]
}